Wrong dependency loop fixed
