+++ cloud/blockstore/libs/storage/core/monitoring_utils.cpp
@@ -4,7 +4,6 @@
 #include "probes.h"
 
 #include <cloud/blockstore/libs/diagnostics/hostname.h>
-#include <cloud/blockstore/libs/storage/partition/model/transaction_time_tracker.h>
 
 #include <cloud/storage/core/libs/common/format.h>
 #include <cloud/storage/core/libs/tablet/blob_id.h>
@@ -52,87 +51,6 @@ TStringBuf AlertClassFromLevel(EAlertLevel alertLevel)
         return "alert-danger";
     default:
         return {};
-    }
-}
-
-void RenderLatencyCell(
-    IOutputStream& out,
-    const TString& trKey,
-    const TString& timeKey)
-{
-    HTML (out) {
-        TABLED () {
-            DIV_CLASS("latency-item")
-            {
-                DIV_CLASS_ID(" ", trKey + "_finished_" + timeKey)
-                {}
-                DIV_CLASS_ID(" ", trKey + "_inflight_" + timeKey)
-                {}
-            }
-        }
-    }
-}
-
-void DumpLatencyForTransactions(
-    IOutputStream& out,
-    const std::span<const TTransactionTimeTracker::TBucketInfo> transactions)
-{
-    HTML (out) {
-        TABLEHEAD () {
-            TABLER () {
-                TABLEH () {
-                    out << "Latency";
-                }
-                for (const auto& trInfo: transactions) {
-                    TABLEH () {
-                        out << trInfo.Description;
-                    }
-                }
-            }
-        }
-
-        for (const auto& [tr, timeKey, timeDescr, timeTooltip]:
-             TTransactionTimeTracker::GetTimeBuckets())
-        {
-            TABLER () {
-                TABLED () {
-                    DIV_CLASS("tooltip-latency")
-                    {
-                        out << timeDescr;
-
-                        if (timeTooltip) {
-                            SPAN_CLASS("tooltiptext-latency")
-                            {
-                                out << timeTooltip;
-                            }
-                        }
-                    }
-                }
-
-                for (const auto& trInfo: transactions) {
-                    RenderLatencyCell(out, trInfo.Key, timeKey);
-                }
-            }
-        }
-    }
-}
-
-void DumpLatencyForTransactions(IOutputStream& out, size_t columnCount)
-{
-    const auto trBuckets = TTransactionTimeTracker::GetTransactionBuckets();
-
-    HTML (out) {
-        TABLE_CLASS("table-latency")
-        {
-            for (size_t i = 0; i < trBuckets.size(); i += columnCount) {
-                const std::size_t chunkSize =
-                    Min(columnCount, trBuckets.size() - i);
-                const std::span<const TTransactionTimeTracker::TBucketInfo>
-                    transactionChunk{&trBuckets[i], chunkSize};
-
-                DumpLatencyForTransactions(out, transactionChunk);
-            }
-        }
     }
 }
 
@@ -1135,106 +1053,6 @@ void DumpTabletNotReady(IOutputStream& out)
     }
 }
 
-void DumpLatency(IOutputStream& out, ui64 tabletId)
-{
-    const TString script = R"(
-        <script>
-            function renderLatency(stat) {
-                for (let key in stat) {
-                    const element = document.getElementById(key);
-                    if (element) {
-                        element.textContent = stat[key];
-                    }
-                }
-            }
-            function loadLatency() {
-                var url = '?action=getLatency';
-                url += '&TabletID=)" +
-                           ToString(tabletId) + R"(';
-                $.ajax({
-                    url: url,
-                    success: function(result) {
-                        renderLatency(result.stat);
-                    },
-                    error: function(jqXHR, status) {
-                        console.log('error');
-                    }
-                });
-            }
-            setInterval(function() { loadLatency(); }, 1000);
-            loadLatency();
-        </script>
-        )";
-
-    const TString style = R"(
-        <style>
-            .table-latency {
-                width: 100%;
-                border-collapse: collapse;
-                padding: 0;
-            }
-            .table-latency th,
-            .table-latency td {
-                padding: 0 8px 0 8px;
-                border: 1px solid black;
-            }
-            .table-latency th {
-                font-weight: bold;
-                text-align: center;
-            }
-            .table-latency td {
-                padding: 0 8px 0 8px;
-                border-top: none;
-                border-bottom: none;
-            }
-            .table-latency tr:last-child {
-                border-bottom: 1px solid black;
-            }
-
-            .tooltip-latency {
-                position: relative;
-                display: inline-block;
-            }
-            .tooltip-latency .tooltiptext-latency {
-                visibility: hidden;
-                width: 120px;
-                background-color: black;
-                color: #fff;
-                text-align: center;
-                border-radius: 6px;
-                padding: 5px 0;
-                position: absolute;
-                z-index: 1;
-            }
-            .tooltip-latency:hover .tooltiptext-latency {
-                visibility: visible;
-            }
-
-            .latency-item {
-                display: flex;
-                width: 100%;
-            }
-            .latency-item > div {
-                flex: 1 1 0;
-                text-align: right;
-            }
-        </style>
-        )";
-
-    HTML (out) {
-        out << script;
-        out << style;
-
-        TAG (TH3) {
-            out << "Transactions";
-        }
-        DumpLatencyForTransactions(out, 9);
-
-        TAG (TH3) {
-            out << "Groups";
-        }
-    }
-}
 
 TCgiParameters GatherHttpParameters(const TEvRemoteHttpInfo& msg)
 {
